Initial commit of the project.
This project aims to provide an application that
will allow a user or group to define dependencies
in their application.  It is tailored towards the
ever-complex EAI (Enterprise Application
Integration) model of an ESB, where many
applications point to many systems.  The process
of managing these connections gets exponentially
complex as integrations are added and removed.

This project aims to at least provide some level
of transparency or additional view that will let
a user identify what applications are touching
which components and services.