Worked on the Contact information, fixing the domain model.
I came across a problem where the one-to-many relationship doesn't
apply for common datatypes.  I couldn't create a hasMany for String
types, for instance, so I just decided to flesh out the model a little
more for that.

I also came across a blog that created a custom
constraint for phone numbers and I used that to create a custom
validator for phone #'s.