Trying to add some JSON-formatted Bootstrap data.
I need to verify if this even works the way I want it to or if this will
end up being more work than it's worth.  I doubt it'll be that much
work, and I think that in the long-run, it might even look better to be
in a JSON format.