Initial addition of the Grails Navigation plugin.
This plugin is pretty slick, and it doesn't take alot of modification on
your controllers to implement.  I also like that it uses the layout/
pages as opposed to having to add to each view component as well.  Very
useful addition.  I'm proposing that this be merged into the trunk
quickly.